use unsigned int for loop indexes

+++ dwl.c
@@ -409,7 +409,7 @@ applyexclusive(struct wlr_box *usable_area,
 	edges[3].negative_axis = &usable_area->width;
 	edges[3].margin = margin_right;
 
-	for (size_t i = 0; i < LENGTH(edges); ++i) {
+	for (unsigned int i = 0; i < LENGTH(edges); ++i) {
 		if ((anchor == edges[i].singular_anchor || anchor == edges[i].anchor_triplet)
 				&& exclusive + edges[i].margin > 0) {
 			if (edges[i].positive_axis)
@@ -585,7 +585,7 @@ arrangelayers(Monitor *m)
 			&usable_area, false);
 
 	// Find topmost keyboard interactive layer, if such a layer exists
-	for (size_t i = 0; i < nlayers; ++i) {
+	for (unsigned int i = 0; i < nlayers; ++i) {
 		wl_list_for_each_reverse(layersurface,
 				&m->layers[layers_above_shell[i]], link) {
 			if (layersurface->layer_surface->current.keyboard_interactive &&
@@ -827,7 +827,7 @@ createmon(struct wl_listener *listener, void *data)
 	m->w = m->m;
 
 	len = LENGTH(m->layers);
-	for (size_t i = 0; i < len; ++i) {
+	for (unsigned int i = 0; i < len; ++i) {
 		wl_list_init(&m->layers[i]);
 	}
 }
